refactor(web): drop 'new-' prefix from src/{api,components,site} (#post-roadmap)
- apps/web/src/new-api → apps/web/src/api
- apps/web/src/new-component → apps/web/src/components (pluralized)
- apps/web/src/new-site → apps/web/src/site

Bulk-rewrites every consumer's front/* alias path:
- front/new-api → front/api
- front/new-component → front/components
- front/new-site → front/site

Also updates apps/web/components.json so shadcn add lands new
components in the renamed paths.

The 'new-' prefix was historical scaffolding (the original Nx React
generator left both old and 'new-' versions). With the modernization
chain landed, the prefix carries no meaning.

- nx affected -t lint test build green (11 tasks)
- grep 'front/new-' apps/web → zero matches outside .next/
