+++ apps/web/app/like/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import LikePage from 'front/new-site/like/LikePage';
+import LikePage from 'front/site/like/LikePage';
 
 export default function Page() {
   return <LikePage />;
