+++ apps/web/app/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import HomePage from 'front/new-site/home';
+import HomePage from 'front/site/home';
 
 export default function Page() {
   return <HomePage />;
