+++ apps/web/app/search/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { SearchPage } from 'front/new-site/search/search';
+import { SearchPage } from 'front/site/search/search';
 
 export default function Page() {
   return <SearchPage />;
