--- apps/web/src/new-api/index.ts
@@ -1,5 +1,0 @@
-export * from 'front/new-api/sido'
-export * from 'front/new-api/sigungu'
-export * from 'front/new-api/shelter'
-export * from 'front/new-api/kind'
-export * from 'front/new-api/animalInfo'
