+++ apps/web/src/site/home/infinity.scss
