+++ apps/web/src/site/search/select/initData.ts
